refactor: Extract scripts' common functionality into a helper module
